Data: split the eliminator construction into Decl.Data
The eliminator machinery had grown 'withDataDecls' well past what fits
in one screen: a single 'bindElims' allocated the fresh binders, built
the surface types of ind-D and rec-D and the path computation rules, and
then elaborated and bound them. The type-building part is pure surface
syntax and does not belong in the monadic checker.

Move it to a new module Rzk.TypeCheck.Decl.Data. That module now owns the
surface-syntax builders (surfaceApps, surfacePi, …), the preprocessed
forms (SortIndex, DataConSurface), and the eliminator construction: an
ElimSpec bundles the declaration's data with the fresh binders and the
validated endpoint images, and 'elimTerms' turns it into the surface
types of the two eliminators and the compute lemmas (an ElimTerms). The
checker allocates the names, validates the endpoints, assembles the
ElimSpec, and elaborates the results as before.

No behavioural change; Decl.hs drops about 255 lines and the eliminator
type-building is now a pure function. The typecheck suite is unaffected.
